Refactor: separate Pool and Pools
